Shifted download IMDF button to original only
