+++ src/pages/conversion/download-imdf.js
@@ -1,11 +1,41 @@
-import { cx } from '@emotion/css';
-import { PrimaryButton } from '@fluentui/react';
-import { failedLogsButton, logsButton } from './style';
+import { cx } from '@emotion/css'; 
+import { PrimaryButton } from '@fluentui/react'; 
+import { logsButton } from './style'; 
+import React, { useState } from 'react';
+import JSZip from 'jszip';
+import { saveAs } from 'file-saver';
 
-export const DownloadIMDF = ({ isFailed, link }) => {
+export const DownloadIMDF = ({ imdfPackageLocation, units, levels, footprint }) => {
+  const [isLoading, setIsLoading] = useState(false);
+
+  const handleUpdateZip = () => {
+    setIsLoading(true);
+    fetch(imdfPackageLocation)
+      .then(response => response.arrayBuffer())
+      .then(data => {
+        const zip = new JSZip();
+        return zip.loadAsync(data);
+      })
+      .then(zip => {
+        zip.file('unit.geojson', JSON.stringify(units, null, 2));
+        zip.file('level.geojson', JSON.stringify(levels, null, 2));
+        zip.file('footprint.geojson', JSON.stringify(footprint, null, 2));
+        return zip.generateAsync({ type: 'blob' });
+      })
+      .then(updatedZip => {
+        saveAs(updatedZip, 'updated_imdf_package.zip');
+        setIsLoading(false);
+      })
+      .catch(() => {
+        setIsLoading(false);
+      });
+  };
+
   return (
-    <a href={link} download target="_blank" rel="noreferrer">
-      <PrimaryButton className={cx(logsButton, { [failedLogsButton]: isFailed })}>Download IMDF</PrimaryButton>
-    </a>
+    <div style={{ cursor: isLoading ? 'wait' : 'default' }}>
+      <PrimaryButton onClick={handleUpdateZip} disabled={isLoading} className={cx(logsButton)}>
+        {isLoading ? 'Downloading..' : 'Download IMDF'}
+      </PrimaryButton>
+    </div>
   );
 };
